chore: move gRPC client module root up
