+++ proto/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./ory/keto/acl/v1alpha1')
